ESC_Backend/requirements.txt
Django==3.2.11
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
PyJWT==2.1.0
psycopg2==2.9.3
python-dateutil==2.8.1
requests==2.25.1
SQLAlchemy==1.4.29
sqlparse==0.4.1
gunicorn==20.1.0
django-cors-headers
